+++ driver/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/bblfsh/sdk/protocol"
+	"github.com/bblfsh/sdk/protocol/driver"
 
 	"github.com/bblfsh/bash-driver/driver/normalizer"
 )
@@ -10,11 +10,11 @@ var version string
 var build string
 
 func main() {
-	d := protocol.Driver{
-		Version:  version,
-		Build:    build,
-		ToNoder:  normalizer.NativeToNoder,
-		Annotate: normalizer.AnnotationRules,
+	d := driver.Driver{
+		Version:          version,
+		Build:            build,
+		ASTParserBuilder: normalizer.ASTParserBuilder,
+		Annotate:         normalizer.AnnotationRules,
 	}
 	d.Exec()
 }
