--- driver/normalizer/normalizer.go
@@ -1,30 +1,0 @@
-package normalizer
-
-import (
-	. "github.com/bblfsh/sdk/uast"
-	. "github.com/bblfsh/sdk/uast/ann"
-
-	"github.com/bblfsh/bash-driver/driver/normalizer/intellij"
-)
-
-const (
-	ErrRootMustBeFile = "root must have internal type FILE"
-)
-
-var AnnotationRules *Rule = On(Any).Self(
-	On(Not(intellij.File)).Error(ErrRootMustBeFile),
-	On(intellij.File).Roles(File).Descendants(
-		On(intellij.Comment).Roles(Comment),
-		On(intellij.Shebang).Roles(Comment, Documentation),
-		// variable declaration
-		On(intellij.VarDefElement).Children(
-			On(intellij.AssignmentWord).Roles(SimpleIdentifier),
-		),
-		// function declaration
-		On(intellij.FunctionDefElement).Children(
-			On(intellij.Function).Roles(FunctionDeclaration),
-			On(intellij.NamedSymbol).Roles(FunctionDeclarationName),
-			On(intellij.GroupElement).Roles(FunctionDeclarationBody, Block),
-		),
-	),
-)
