+++ driver/normalizer/parser_test.go
@@ -1,0 +1,18 @@
+package normalizer
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestToNoderVeryLong(t *testing.T) {
+	require := require.New(t)
+
+	f, err := getFixture(integration, "very_long.native")
+	require.NoError(err)
+
+	n, err := ToNoder.ToNode(f)
+	require.NoError(err)
+	require.NotNil(n)
+}
