+++ tests/comments.native
@@ -1,4 +1,6 @@
 {
+    "status": "ok",
+    "errors": [],
     "ast": {
         "children": [
             {
@@ -180,7 +182,5 @@
         "startOffset": 0,
         "text": "# comment 1\n\na=\"# not a comment\"\n\nb=\"foo bar\" # comment 2\n",
         "textLength": 58
-    },
-    "errors": [],
-    "status": "ok"
+    }
 }
