+++ tests/var_declaration.uast
@@ -5,25 +5,22 @@ FILE {
 .  Roles: File
 .  TOKEN "a=1
 "
+.  StartPosition: {
+.  .  Offset: 0
+.  .  Line: 0
+.  .  Col: 0
+.  }
 .  Properties: {
 .  .  textLength: 4
 .  }
 .  Children: {
-.  .  0: [Bash] linefeed {
-.  .  .  TOKEN "
-"
+.  .  0: simple-command {
+.  .  .  TOKEN "a=1"
 .  .  .  StartPosition: {
-.  .  .  .  Offset: 3
+.  .  .  .  Offset: 0
 .  .  .  .  Line: 0
 .  .  .  .  Col: 0
 .  .  .  }
-.  .  .  Properties: {
-.  .  .  .  internalRole: children
-.  .  .  .  textLength: 1
-.  .  .  }
-.  .  }
-.  .  1: simple-command {
-.  .  .  TOKEN "a=1"
 .  .  .  Properties: {
 .  .  .  .  internalRole: children
 .  .  .  .  textLength: 3
@@ -31,12 +28,30 @@ FILE {
 .  .  .  Children: {
 .  .  .  .  0: var-def-element {
 .  .  .  .  .  TOKEN "a=1"
+.  .  .  .  .  StartPosition: {
+.  .  .  .  .  .  Offset: 0
+.  .  .  .  .  .  Line: 0
+.  .  .  .  .  .  Col: 0
+.  .  .  .  .  }
 .  .  .  .  .  Properties: {
 .  .  .  .  .  .  internalRole: children
 .  .  .  .  .  .  textLength: 3
 .  .  .  .  .  }
 .  .  .  .  .  Children: {
-.  .  .  .  .  .  0: [Bash] = {
+.  .  .  .  .  .  0: [Bash] assignment_word {
+.  .  .  .  .  .  .  Roles: SimpleIdentifier
+.  .  .  .  .  .  .  TOKEN "a"
+.  .  .  .  .  .  .  StartPosition: {
+.  .  .  .  .  .  .  .  Offset: 0
+.  .  .  .  .  .  .  .  Line: 0
+.  .  .  .  .  .  .  .  Col: 0
+.  .  .  .  .  .  .  }
+.  .  .  .  .  .  .  Properties: {
+.  .  .  .  .  .  .  .  internalRole: children
+.  .  .  .  .  .  .  .  textLength: 1
+.  .  .  .  .  .  .  }
+.  .  .  .  .  .  }
+.  .  .  .  .  .  1: [Bash] = {
 .  .  .  .  .  .  .  TOKEN "="
 .  .  .  .  .  .  .  StartPosition: {
 .  .  .  .  .  .  .  .  Offset: 1
@@ -48,7 +63,7 @@ FILE {
 .  .  .  .  .  .  .  .  textLength: 1
 .  .  .  .  .  .  .  }
 .  .  .  .  .  .  }
-.  .  .  .  .  .  1: [Bash] combined word {
+.  .  .  .  .  .  2: [Bash] combined word {
 .  .  .  .  .  .  .  TOKEN "1"
 .  .  .  .  .  .  .  StartPosition: {
 .  .  .  .  .  .  .  .  Offset: 2
@@ -74,16 +89,21 @@ FILE {
 .  .  .  .  .  .  .  .  }
 .  .  .  .  .  .  .  }
 .  .  .  .  .  .  }
-.  .  .  .  .  .  2: [Bash] assignment_word {
-.  .  .  .  .  .  .  Roles: SimpleIdentifier
-.  .  .  .  .  .  .  TOKEN "a"
-.  .  .  .  .  .  .  Properties: {
-.  .  .  .  .  .  .  .  internalRole: children
-.  .  .  .  .  .  .  .  textLength: 1
-.  .  .  .  .  .  .  }
-.  .  .  .  .  .  }
 .  .  .  .  .  }
 .  .  .  .  }
+.  .  .  }
+.  .  }
+.  .  1: [Bash] linefeed {
+.  .  .  TOKEN "
+"
+.  .  .  StartPosition: {
+.  .  .  .  Offset: 3
+.  .  .  .  Line: 0
+.  .  .  .  Col: 0
+.  .  .  }
+.  .  .  Properties: {
+.  .  .  .  internalRole: children
+.  .  .  .  textLength: 1
 .  .  .  }
 .  .  }
 .  }
