+++ tests/var_expansion.native
@@ -1,4 +1,6 @@
 {
+    "status": "ok",
+    "errors": [],
     "ast": {
         "children": [
             {
@@ -292,7 +294,5 @@
         "startOffset": 0,
         "text": "$A\n${B}\n\"$C\"\n\"${D}\"\n",
         "textLength": 20
-    },
-    "errors": [],
-    "status": "ok"
+    }
 }
